+++ algoritmo1/src/main/java/usac/repositories/MovimientoRepository.java
@@ -4,7 +4,6 @@
  */
 package usac.repositories;
 
-import java.sql.Connection;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
 import java.sql.SQLException;
@@ -17,7 +16,7 @@ import usac.db.Conexion;
 public class MovimientoRepository {
 
     // Método para establecer el valor inicial
-    public void setInitialValue(int value) throws SQLException {
+    public void setInitialValue(double value) throws SQLException {
 
         // Comprobar si ya existe un registro
         String checkSql = "SELECT COUNT(*) FROM Movimiento";
@@ -27,14 +26,14 @@ public class MovimientoRepository {
                 // Si existe, actualizar el valor
                 String updateSql = "UPDATE Movimiento SET valor = ? WHERE id = 1";
                 try (PreparedStatement updateStmt = Conexion.conexion.prepareStatement(updateSql)) {
-                    updateStmt.setInt(1, value);
+                    updateStmt.setDouble(1, value);
                     updateStmt.executeUpdate();
                 }
             } else {
                 // Si no existe, insertar un nuevo registro
                 String insertSql = "INSERT INTO Movimiento (id, valor) VALUES (1, ?)";
                 try (PreparedStatement insertStmt = Conexion.conexion.prepareStatement(insertSql)) {
-                    insertStmt.setInt(1, value);
+                    insertStmt.setDouble(1, value);
                     insertStmt.executeUpdate();
                 }
             }
@@ -42,20 +41,20 @@ public class MovimientoRepository {
 
     }
 
-    public int getValue() throws SQLException {
+    public double getValue() throws SQLException {
         String sql = "SELECT valor FROM Movimiento WHERE id = 1";
         try (PreparedStatement stmt = Conexion.conexion.prepareStatement(sql); ResultSet rs = stmt.executeQuery()) {
             if (rs.next()) {
-                return rs.getInt("valor");
+                return rs.getDouble("valor");
             }
         }
         return 0;
     }
 
-    public boolean incrementValue(int increment) throws SQLException {
+    public boolean incrementValue(double increment) throws SQLException {
         String sql = "UPDATE Movimiento SET valor = valor + ? WHERE id = 1";
         try (PreparedStatement stmt = Conexion.conexion.prepareStatement(sql)) {
-            stmt.setInt(1, increment);
+            stmt.setDouble(1, increment);
             int executeUpdate = stmt.executeUpdate();
             if (executeUpdate >= 1) {
                 return true;
@@ -64,10 +63,10 @@ public class MovimientoRepository {
         return false;
     }
 
-    public boolean decrementValue(int decrement) throws SQLException {
+    public boolean decrementValue(double decrement) throws SQLException {
         String sql = "UPDATE Movimiento SET valor = valor - ? WHERE id = 1";
         try (PreparedStatement stmt = Conexion.conexion.prepareStatement(sql)) {
-            stmt.setInt(1, decrement);
+            stmt.setDouble(1, decrement);
             int executeUpdate = stmt.executeUpdate();
             if (executeUpdate >= 1) {
                 return true;
